Refactor code / fix errors
refactor the main loop (as in the actual imagemagick command) into many calls to imagemagick
this makes adding new features easier.
Also we made handling of dimentions a bit better. (although it might still be a problem if the aspect ratios and the area of the files are wonky / same)
there appears to be no regressions but the old script is also uploaded if the need arises.